Subject: Implement majority of chunk_type
Problem:
Needed to implement rest of chunk type spec

Solution:
Followed spec to implement rest of chunk type spec

Note:
The is_valid check may be wrong as the way im constructing the error
types on the init of the chunk type causes 2 tests to panic. I tried to
override the is_err method but that seemed to do nothing. Will try to
resolve this tomorrow